IncomeCurrentWH

+++ src/Datiss.Budget.ViewModels/IncomeCurrentWHViewModel.cs
@@ -29,25 +29,25 @@ namespace Datiss.Budget.ViewModels
 
         public int ConsumptionUser { get; set; }
 
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
 
-        public int Income { get; set; }
+        public long Income { get; set; }
 
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
 
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
 
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
 
-        public int TIncome { get; set; }
+        public long TIncome { get; set; }
 
         public int Diff_ConsWsVolume { get; set; }
 
-        public int Note2Income { get; set; }
+        public long Note2Income { get; set; }
 
-        public int WasteVolume { get; set; }
+        public long WasteVolume { get; set; }
 
     }
 
@@ -91,31 +91,31 @@ namespace Datiss.Budget.ViewModels
         public string ConsumptionUserDisplay => ConsumptionUser.ToString("N0");
 
 
-        public int Cost { get; set; }
+        public long Cost { get; set; }
         public string CostDisplay => Cost.ToString("N0");
 
 
-        public int Note3Price { get; set; }
+        public long Note3Price { get; set; }
         public string Note3PriceDisplay => Note3Price.ToString("N0");
 
 
-        public int Income { get; set; }
+        public long Income { get; set; }
         public string IncomeDisplay => Income.ToString("N0");
 
 
-        public int Note3Income { get; set; }
+        public long Note3Income { get; set; }
         public string Note3IncomeDisplay => Note3Income.ToString("N0");
 
 
-        public int SubscriptionIncome { get; set; }
+        public long SubscriptionIncome { get; set; }
         public string SubscriptionIncomeDisplay => SubscriptionIncome.ToString("N0");
 
 
-        public int SeasonalIncome { get; set; }
+        public long SeasonalIncome { get; set; }
         public string SeasonalIncomeDisplay => SeasonalIncome.ToString("N0");
 
 
-        public int TIncome { get; set; }
+        public long TIncome { get; set; }
         public string TIncomeDisplay => TIncome.ToString("N0");
